backend/requirements.txt
Django>=5.2.5
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
google-generativeai>=0.3.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
